pyproject.toml
[project]
name = "chess-tutor"
version = "0.1.0"
description = "Interactive chess tutor with AI-powered analysis"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.12.0",
    "python-chess>=1.10.0",
    "stockfish>=3.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
chess-tutor = "examples.demo:main"

[tool.uv]
dev-dependencies = []
